[fix] frontend voltage is 10 bits, so maximum 1023
The comments were correctly saying it's 10 bits... but incorrectly
translating it as maximum 4095. That's actually 1023.

It doesn't have much effect on the software behaviour, as if the user
passed a value > 1023, the device would just complain, and the command
would fail pretty much the same way as with detecting it in Python.